array_prep test with hypoyhesis + comments
Correction of array_prep test n1 without hypothesis, test with randn (just to play) and with hypothesis (everything in the comments).  Still some problem that are listed in the code...
Correction of polyfit_global (previously I was using array_prep_v1!!) + what happened  in the fitting procedure if the x data are not ordered. Some comment on array_prep_v2 regarding the sign of uncertainty.